Add victory screen, Remove transition wait time Finish Tutorial Video P15 - COMPLETED Tutorial
Upon finishing all levels a button to the victory
screen appears allowing players to navigate to the
victory screen. On the victory screen it displays
"You Win" and a button to return to the main menu.

Removed the await for the fade from black meaning that
the player only has to wait for the countdown and not
the transition and countdown/